address comments 2
