+++ repl/test/dotty/tools/repl/LazyValsWarningTest.scala
@@ -1,5 +1,11 @@
 package dotty.tools.repl
 
+import java.io.File
+import java.nio.charset.StandardCharsets.UTF_8
+
+import scala.io.Source
+import scala.util.Using
+
 import org.junit.Assert.{assertFalse, assertTrue}
 import org.junit.Assume.assumeTrue
 import org.junit.Test
@@ -16,13 +22,14 @@ import org.junit.Test
 class LazyValsWarningTest:
 
   @Test def noUnsafeLazyValsWarning(): Unit =
-    val javaHomeOverride = ReplTestProcess.javaHomeOverride
+    val javaHomeOverride =
+      sys.props.get("dotty.tests.replJavaHome").orElse(sys.env.get("DOTTY_REPL_TEST_JAVA_HOME"))
     assumeTrue(
       "warning is only emitted on JDK 24+",
       javaHomeOverride.isDefined || Runtime.version.feature >= 24
     )
 
-    val output = ReplTestProcess.output("val replCheck = 1 + 1")
+    val output = runReplInitScript(javaHomeOverride, "val replCheck = 1 + 1")
 
     assertTrue(
       s"REPL did not evaluate the init script, output was:\n$output",
@@ -33,5 +40,27 @@ class LazyValsWarningTest:
         s"REPL emitted a legacy lazy vals warning ($marker):\n$output",
         output.contains(marker)
       )
+
+  /** Runs `initScript` in a freshly spawned REPL subprocess and returns its
+   *  merged stdout/stderr output.
+   */
+  private def runReplInitScript(javaHomeOverride: Option[String], initScript: String): String =
+    val javaHome = javaHomeOverride.getOrElse(sys.props("java.home"))
+    val javaBin = List(javaHome, "bin", "java").mkString(File.separator)
+    val classpath = sys.props("java.class.path")
+    val command = List(
+      javaBin,
+      "-Dscala.usejavacp=true",
+      "-classpath", classpath,
+      "dotty.tools.repl.Main",
+      "-usejavacp",
+      "-color:never",
+      "--repl-quit-after-init",
+      "--repl-init-script", initScript,
+    )
+    val process = ProcessBuilder(command*).redirectErrorStream(true).start()
+    val output = Using.resource(Source.fromInputStream(process.getInputStream, UTF_8.name))(_.mkString)
+    process.waitFor()
+    output
 
 end LazyValsWarningTest
